requirements.txt
streamlit>=1.28.0
mediapipe>=0.10.21
opencv-python-headless>=4.8.0
numpy>=1.21.0,<2.0.0
pillow>=8.0.0
requests>=2.25.0